Add ROS publishers for all classifiers
config/emobase_live4_ros_fixed.conf serves as the template configuration to show
how one can publish OpenSmile messages to ROS.

The topic is configured in the conf file and the message format is an opensmile
message encoded in a JSON string. The meaning of the contents of the Opensmile
message vary from component to component.

Signed-off-by: George Paraskevopoulos <George Paraskevopoulos>